+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add rule extension point [#12](https://github.com/nbbrd/heylogs/issues/12)
 - Add failure format extension point [#13](https://github.com/nbbrd/heylogs/issues/13)
+- Add unreleased pattern to version filtering [#19](https://github.com/nbbrd/heylogs/issues/19)
 
 ### Changed